refactoring "all values"
