+++ mod/01_rating/set/abstract/import.rb
