+++ mod/01_rating/set/type/metric.rb
@@ -9,6 +9,7 @@ card_accessor :value_type
 card_accessor :report_type
 card_accessor :research_policy
 card_accessor :project
+card_accessor :all_values
 
 def metric_type
   metric_type_card.item_names.first
