+++ mod/01_rating/spec/set/type_plus_right/metric/value_type_spec.rb
