+++ mod/01_rating/spec/set/type_plus_right/metric_value/source_spec.rb
