--- mod/wikirate/set/type_plus_right/metric/all_values.rb
@@ -1,7 +1,0 @@
-include_set Abstract::AllValues
-
-format do
-  def page_link_params
-    [:name, :industry, :project, :year, :value]
-  end
-end
